ai_model/requirements.txt
absl-py==2.4.0
astunparse==1.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.1
h5py==3.15.1
idna==3.11
joblib==1.5.3
keras==3.13.2
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.8.2
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==1.26.3
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pandas==2.2.0
pillow==12.1.0
protobuf==5.29.6
psutil==6.0.0
psycopg2-binary==2.9.9
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.5
rich==14.3.3
scikit-learn==1.4.0
scipy==1.16.3
setuptools==82.0.0
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
termcolor==3.3.0
threadpoolctl==3.6.0
tqdm==4.66.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.1
